Improved findCliques and fitness functions
I also moved bees over to evolution and will be working on it from there. Modified the dna format to be the familiar DNA bit format as opposed to the flattened matrix -- the advantage of this is that I have a recursive formula to mutate edges and such within the DNA bit in this current format as opposed to the old format which was just the matrix flattened.